RebindAllTheKeys Compat closes #50 (again)
